+++ code/CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#809](https://github.com/InditexTech/weavejs/issues/806) Center alignment guides during element drag for a frame in Weave canvas
 
+### Changed
+
+- [#808](https://github.com/InditexTech/weavejs/issues/808) Homologate Websockets store to load initial room state as the Azure Web PubSub does
+
 ### Fixed
 
 - [#806](https://github.com/InditexTech/weavejs/issues/806) Line size handlers not positioned correctly inside a frame